requirements.txt
alabaster==0.7.12
asgiref==3.4.1
Babel==2.9.1
certifi==2021.5.30
charset-normalizer==2.0.4
colorama==0.4.4
coverage==5.5
Django==3.2.6
djangorestframework==3.12.4
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
factory-boy==3.2.0
Faker==8.11.0
fett==0.3.2
idna==3.2
imagesize==1.2.0
Jinja2==3.0.1
livereload==2.6.3
MarkupSafe==2.0.1
networkx==2.6.2
packaging==21.0
Pygments==2.9.0
pymongo==3.11.4
pyparsing==2.4.7
python-dateutil==2.8.2
python-jsonrpc-server==0.3.4
pytz==2021.1
PyYAML==5.4.1
requests==2.26.0
rstcheck==3.3.1
six==1.16.0
snooty-lextudio==1.11.0.dev0
snowballstemmer==2.1.0
Sphinx==4.1.2
sphinx-autobuild==2021.3.14
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
tornado==6.1
typing-extensions==3.10.0.0
ujson==1.35
urllib3==1.26.6
watchdog==1.0.2
